# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Paperless Archivist API
  version: 1.18.1
servers:
  - url: /
security:
  - cookieSession: []
  - bearerToken: []
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: pa_session
    bearerToken:
      type: http
      scheme: bearer
    metricsToken:
      type: http
      scheme: bearer
      description: Dedicated bearer token configured through ARCHIVIST_METRICS_TOKEN; application API tokens are not accepted.
    webhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Shared secret configured through ARCHIVIST_WEBHOOK_SECRET.
  responses:
    BadRequest:
      description: Invalid path, query, or request payload
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    Stage:
      type: string
      enum: [ocr, metadata, apply]
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    MeResponse:
      type: object
      required: [username, roles, permissions]
      properties:
        username: { type: string }
        roles:
          type: array
          items: { $ref: "#/components/schemas/Role" }
        permissions: { $ref: "#/components/schemas/PermissionFlags" }
        csrf_token:
          type: [string, "null"]
...
    Session:
      type: object
      additionalProperties: true
...
paths:
...
  /api/auth/login:
    post:
      x-archivist-permission: public
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
  /api/auth/paperless-login:
    post:
      x-archivist-permission: public
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in through the optional Paperless login bridge
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
...
  /api/auth/oidc/login:
    get:
      x-archivist-permission: public
      security: []
      description: >-
        Redirects to the OIDC provider and sets the short-lived HttpOnly
        pa_oidc_state cookie that binds the login attempt to this browser.
      parameters:
        - name: return_to
          in: query
          schema: { type: string }
      responses:
        "307": { description: Redirect to OIDC provider }
  /api/auth/oidc/callback:
    get:
      x-archivist-permission: public
      security: []
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
        - name: error
          in: query
          schema: { type: string }
        - name: error_description
          in: query
          schema: { type: string }
      description: >-
        The state parameter must match the pa_oidc_state cookie set by
        /api/auth/oidc/login in the same browser; the cookie is cleared on
        every callback outcome.
      responses:
        "303": { description: Login completed and session cookies issued }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/webhooks/paperless/document-consumed:
    post:
      x-archivist-permission: public
      summary: Queue newly consumed Paperless documents immediately
      description: >-
        Machine-to-machine endpoint outside user authentication. The shared
        X-Webhook-Secret is compared against ARCHIVIST_WEBHOOK_SECRET.
      security:
        - webhookSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WebhookConsumedRequest" }
      responses:
        "202":
          description: Documents accepted and queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueueCountResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "500": { $ref: "#/components/responses/InternalServerError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /api/auth/change-password:
    post:
      x-archivist-permission: authenticated
      security:
        - cookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password }
      responses:
        "200": { description: Password changed and sessions revoked }
  /api/auth/sessions:
    get:
      x-archivist-permission: authenticated
      description: List browser sessions. Requires an interactive cookie session; bearer tokens are always rejected regardless of scope.
      security:
        - cookieSession: []
      responses:
        "200":
          description: Own sessions, or all sessions for an administrator
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Session" }
        "403": { description: Interactive cookie session required }
  /api/auth/sessions/{id}/revoke:
    post:
      x-archivist-permission: manage_users
      description: Revoke a browser session as an administrator. Requires an interactive cookie session; bearer tokens are always rejected regardless of scope.
      security:
        - cookieSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Session revoked }
        "403": { description: Interactive administrator cookie session required }
  /api/settings:
    get:
      x-archivist-permission: read_settings
      responses:
        "200":
          description: Runtime settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeSettings" }
    put:
      x-archivist-permission: write_settings
      security:
        - cookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSettingsRequest" }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeSettings" }
...
  /api/notifications/test:
    post:
      x-archivist-permission: write_settings
      security:
        - cookieSession: []
      responses:
        "200":
          description: Notification webhook delivery test result
  /api/model-providers/test:
    post:
      x-archivist-permission: write_settings
      security:
        - cookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProviderTestRequest" }
      responses:
        "200":
          description: Provider connection result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderTestResponse" }
  /api/model-providers/{name}/models:
    post:
      x-archivist-permission: read_settings
      security:
        - cookieSession: []
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Installed Ollama models for a configured provider
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OllamaInstalledModelsResponse" }
...
  /api/prompts:
    get:
      x-archivist-permission: read_settings
      responses:
        "200":
          description: Prompts
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Prompt" }
    post:
      x-archivist-permission: write_settings
      security:
        - cookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stage, name, content]
              properties:
                stage: { $ref: "#/components/schemas/Stage" }
                name: { type: string }
                content: { type: string }
                output_schema: true
                activate: { type: boolean }
      responses:
        "200": { description: Created prompt }
...
  /api/prompts/experiments:
    get:
      x-archivist-permission: read_settings
      summary: Aggregate prompt experiment review outcomes
      description: Requires the settings:read permission.
      responses:
        "200":
          description: Prompt experiment outcome groups
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PromptExperimentsResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /api/prompts/{id}/activate:
    post:
      x-archivist-permission: write_settings
      security:
        - cookieSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Prompt activated }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/prompts/test:
    post:
      x-archivist-permission: write_settings
      security:
        - cookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TestPromptRequest" }
      responses:
        "200":
          description: Prompt test result with parsed output and validation errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestPromptResponse" }
        "409": { $ref: "#/components/responses/NotConfigured" }
  /api/paperless/sync-metadata:
    post:
      x-archivist-permission: write_batches
      responses:
        "200":
          description: Sync summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaperlessSyncResponse" }
        "409": { $ref: "#/components/responses/NotConfigured" }
  /api/paperless/consistency:
    get:
      x-archivist-permission: read_inventory
      responses:
        "200":
          description: Paperless metadata consistency report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaperlessConsistencyResponse" }
        "409": { $ref: "#/components/responses/NotConfigured" }
  /api/paperless/completion-tags/reconcile:
    post:
      x-archivist-permission: write_batches
      description: >-
        Dry-run or apply the global completion tag when every enabled stage is
        proven terminal either by the complete set of Paperless stage tags or
        by the complete set of local inventory statuses. Active and
        review-waiting runs are excluded; status-based writes are rechecked
        under the per-document run lock. Pass document_ids from a dry-run plan
        to pin the apply set.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReconcileCompletionTagsRequest" }
      responses:
        "200":
          description: Completion tag reconciliation plan or applied changes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReconcileCompletionTagsResponse" }
        "409": { $ref: "#/components/responses/NotConfigured" }
  /api/dashboard:
    get:
      x-archivist-permission: read_dashboard
      parameters:
        - name: range
          in: query
          schema:
            type: string
            enum: ["24h", "7d", "30d", "90d", "12m", "all"]
            default: "24h"
      responses:
        "200":
          description: Dashboard counts and analytics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DashboardResponse" }
...
  /api/statistics:
    get:
      x-archivist-permission: read_dashboard
      description: >-
        Usage, cost and pipeline-throughput analytics over a custom time range
        for the Statistics page. Defaults apply only when a parameter is
        absent or blank; a present-but-unparseable value is rejected with 400.
      parameters:
        - name: from
          in: query
          description: >-
            Inclusive range start: an RFC3339 timestamp or a bare YYYY-MM-DD
            date (the day's first instant, UTC). Defaults to `to` minus 30
            days.
          schema: { type: string }
        - name: to
          in: query
          description: >-
            Exclusive range end: an RFC3339 timestamp or a bare YYYY-MM-DD
            date. A bare date means the END of that day (the next UTC
            midnight), so `from` and `to` naming the same day cover exactly
            that day. Defaults to now.
          schema: { type: string }
        - name: bucket
          in: query
          description: Time-series bucket granularity.
          schema:
            type: string
            enum: ["hour", "day", "week", "month"]
            default: "day"
      responses:
        "200":
          description: Statistics summary, zero-filled time series and breakdowns
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatisticsResponse" }
        "400":
          description: Unparseable from/to, from not before to, or unknown bucket
  /api/workflow/mode:
    put:
      x-archivist-permission: write_settings
      security:
        - cookieSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateWorkflowModeRequest" }
      responses:
        "200":
          description: Updated runtime settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeSettings" }
...
  /api/reviews:
    get:
      x-archivist-permission: read_reviews
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        "200": { description: Review queue }
...